requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
qdrant-client>=1.12.0
openai>=1.50.0
langfuse>=2.50.0
python-dotenv>=1.0.0
pypdf>=5.0.0
requests>=2.32.0
tiktoken>=0.8.0
